+++ VacSuit.java
@@ -1,10 +1,10 @@
 public class VacSuit extends Armor {
 
-  public static final int VACSUIT_WEIGHT = 50;
+  public static final double VACSUIT_WEIGHT = 50.0;
   public static final int VACSUIT_ATTACK = 1;
   public static final int VACSUIT_DEFENSE = 3;
   
-  public VacSuit(int weight, int attack, int defense) {
+  public VacSuit(double weight, int attack, int defense) {
     super(VACSUIT_WEIGHT, VACSUIT_ATTACK, VACSUIT_DEFENSE);
   }
